+++ R/statistics.R
@@ -338,7 +338,6 @@ get_skewness <- function(x, narm = TRUE) {
       x <- x[!ina]
   }
 
-
   n <- length(x)
   if(n < 3)
     stop("Skewness requires at least 3 complete observations.")
@@ -347,36 +346,8 @@ get_skewness <- function(x, narm = TRUE) {
   y <- sqrt(n) * sum(x ^ 3, na.rm = narm) / (sum(x ^ 2, na.rm = narm) ^ (3/2))
   ret <- y * sqrt(n * (n - 1)) / (n - 2)
 
-
   return(ret)
-
 }
-
-# skewness <-
-#   function(x, na.rm = FALSE, type = 3)
-#   {
-#     if(any(ina <- is.na(x))) {
-#       if(na.rm)
-#         x <- x[!ina]
-#       else
-#         return(NA)
-#     }
-#
-#     if(!(type %in% (1 : 3)))
-#       stop("Invalid 'type' argument.")
-#
-#     n <- length(x)
-#     x <- x - mean(x)
-#     y <- sqrt(n) * sum(x ^ 3) / (sum(x ^ 2) ^ (3/2))
-#     if(type == 2) {
-#       if(n < 3)
-#         stop("Need at least 3 complete observations.")
-#       y <- y * sqrt(n * (n - 1)) / (n - 2)
-#     } else if(type == 3)
-#       y <- y * ((1 - 1 / n)) ^ (3/2)
-#
-#     y
-#   }
 
 
 get_kurtosis <- function(x, narm = TRUE) {
@@ -401,34 +372,7 @@ get_kurtosis <- function(x, narm = TRUE) {
   return(ret)
 }
 
-# kurtosis <-
-#   function(x, na.rm = FALSE, type = 3)
-#   {
-#     if(any(ina <- is.na(x))) {
-#       if(na.rm)
-#         x <- x[!ina]
-#       else
-#         return(NA)
-#     }
-#
-#     if(!(type %in% (1 : 3)))
-#       stop("Invalid 'type' argument.")
-#
-#     n <- length(x)
-#     x <- x - mean(x)
-#     r <- n * sum(x ^ 4) / (sum(x ^ 2) ^ 2)
-#     y <- if(type == 1)
-#       r - 3
-#     else if(type == 2) {
-#       if(n < 4)
-#         stop("Need at least 4 complete observations.")
-#       ((n + 1) * (r - 3) + 6) * (n - 1) / ((n - 2) * (n - 3))
-#     }
-#     else
-#       r * (1 - 1 / n) ^ 2 - 3
-#
-#     y
-#   }
+
 
 
 # get_cmh <- function(x, y, wgt = NULL, corrct = FALSE, bylbl = "", output = FALSE) {